# docs/architecture/api_specification.yaml
openapi: 3.0.3
info:
  title: Portfolio Optimization API
  description: |
    REST API for the Portfolio Optimization Platform.

    ## Overview
    This API provides access to cryptocurrency price data and portfolio optimization results.

    ## Authentication
    **MVP:** No authentication required (internal use only)

    **Production:** API key authentication via `X-API-Key` header (to be implemented)

    ## Rate Limits
    **MVP:** No rate limits

    **Production:** 100 requests/minute per IP (to be implemented)

  version: 1.0.0
  contact:
    name: Data Engineering Team
    email: data-team@company.com
  license:
    name: Internal Use Only
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.portfolio.company.com
    description: Production (future)
...
paths:
...
  /klines/{symbol}:
    get:
      tags:
        - Data
      summary: Get Price History
      description: Returns OHLCV (Open, High, Low, Close, Volume) price history for a symbol
      operationId: getKlines
      parameters:
        - name: symbol
          in: path
          required: true
          description: Trading pair symbol (e.g., BTCUSDT)
          schema:
            type: string
            example: BTCUSDT
      responses:
        '200':
          description: Price history data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KlinesResponse'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /portfolio:
    get:
      tags:
        - Analytics
      summary: Get Optimal Portfolio
      description: |
        Returns optimized portfolio weights using Markowitz mean-variance optimization.

        The optimization maximizes Sharpe ratio subject to:
        - Sum of weights = 1 (fully invested)
        - All weights >= 0 (no short selling)
      operationId: getPortfolio
      responses:
        '200':
          description: Optimal portfolio allocation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
              example:
                weights:
                  BTCUSDT: 0.352
                  ETHUSDT: 0.248
                  BNBUSDT: 0.185
                  SOLUSDT: 0.123
                  ADAUSDT: 0.092
                expected_return: 0.185
                volatility: 0.283
                sharpe_ratio: 1.24
                optimization_method: "scipy"
                computed_at: "2025-01-15T00:05:32.123Z"
        '404':
          description: No portfolio computed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/summary:
    get:
      tags:
        - Analytics
      summary: Get Portfolio Summary
      description: |
        Returns a lightweight summary of the optimal portfolio with
        weights, expected return, volatility, and Sharpe ratio.
      operationId: getPortfolioSummary
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummaryResponse'
        '404':
          description: No portfolio computed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/frontier:
    get:
      tags:
        - Analytics
      summary: Get Efficient Frontier
      description: |
        Returns the efficient frontier showing optimal risk-return tradeoffs.

        Includes the frontier curve, max Sharpe portfolio, min variance portfolio,
        individual asset positions, and the Capital Market Line.
      operationId: getEfficientFrontier
      responses:
        '200':
          description: Efficient frontier data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontierResponse'
        '404':
          description: Frontier not yet computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /portfolio/backtest:
    get:
      tags:
        - Analytics
      summary: Get Backtest Results
      description: |
        Returns walk-forward backtest results comparing the optimized strategy
        against equal-weight and BTC-only benchmarks.

        Uses rolling windows (train/test split) to validate out-of-sample performance.
      operationId: getBacktestResults
      responses:
        '200':
          description: Backtest results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
        '404':
          description: Backtest not yet computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/monte-carlo:
    get:
      tags:
        - Analytics
      summary: Get Monte Carlo Simulation
      description: |
        Returns Monte Carlo simulation results with percentile bands,
        final value distribution, and risk metrics (VaR, CVaR).
      operationId: getMonteCarlo
      responses:
        '200':
          description: Monte Carlo simulation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonteCarloResponse'
        '404':
          description: Simulation not yet computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/combined:
    get:
      tags:
        - Analytics
      summary: Get Combined Portfolio
      description: |
        Returns a blended crypto + traditional portfolio with configurable
        allocation split. Default is 60% crypto / 40% traditional.
      operationId: getCombinedPortfolio
      parameters:
        - name: crypto_weight
          in: query
          required: false
          description: Fraction allocated to crypto (0-1). Default 0.6.
          schema:
            type: number
            format: double
            default: 0.6
            minimum: 0
            maximum: 1
      responses:
        '200':
          description: Combined portfolio allocation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CombinedPortfolioResponse'
        '404':
          description: One or both sub-portfolios not computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/rebalance:
    get:
      tags:
        - Analytics
      summary: Get Rebalancing Alerts
      description: |
        Compares current portfolio weights against optimal allocation and
        flags symbols that have drifted beyond the threshold. Suggests trades.
      operationId: getRebalanceAlerts
      parameters:
        - name: drift_threshold
          in: query
          required: false
          schema:
            type: number
            format: double
            default: 0.05
        - name: portfolio_value
          in: query
          required: false
          schema:
            type: number
            format: double
            default: 10000.0
        - name: portfolio_key
          in: query
          required: false
          schema:
            type: string
            default: weights
      responses:
        '200':
          description: Rebalancing alerts and suggested trades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebalanceResponse'
        '404':
          description: Optimal portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/trad:
    get:
      tags:
        - Analytics
      summary: Get Traditional Asset Portfolio
      description: |
        Returns optimized portfolio weights for traditional assets (stocks, ETFs,
        commodities) using Markowitz mean-variance optimization with 252 trading days.
      operationId: getTradPortfolio
      responses:
        '200':
          description: Traditional portfolio allocation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
        '404':
          description: Traditional portfolio not computed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/trad/frontier:
    get:
      tags:
        - Analytics
      summary: Get Traditional Asset Efficient Frontier
      description: |
        Returns the efficient frontier for traditional assets showing optimal
        risk-return tradeoffs with 252 trading days per year.
      operationId: getTradFrontier
      responses:
        '200':
          description: Traditional efficient frontier data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrontierResponse'
        '404':
          description: Traditional frontier not yet computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /portfolio/trad/backtest:
    get:
      tags:
        - Analytics
      summary: Get Traditional Asset Backtest Results
      description: |
        Returns walk-forward backtest results for traditional assets comparing
        the optimized strategy against equal-weight and SPY-only benchmarks.
      operationId: getTradBacktest
      responses:
        '200':
          description: Traditional backtest results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
        '404':
          description: Traditional backtest not yet computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    KlineRecord:
      type: object
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
      properties:
        timestamp:
          type: string
          description: Candle open timestamp
        open:
          type: number
          format: double
          description: Opening price
        high:
          type: number
          format: double
          description: Highest price
        low:
          type: number
          format: double
          description: Lowest price
        close:
          type: number
          format: double
          description: Closing price
        volume:
          type: number
          format: double
          description: Trading volume
    KlinesResponse:
      type: object
      required:
        - symbol
        - count
        - data
      properties:
        symbol:
          type: string
          description: Trading pair symbol
        count:
          type: integer
          description: Number of kline records
        data:
          type: array
          items:
            $ref: '#/components/schemas/KlineRecord'
...
    PortfolioResponse:
      type: object
      required:
        - weights
        - expected_return
        - volatility
        - sharpe_ratio
      properties:
        weights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Optimal weight for each symbol (sums to 1.0)
        expected_return:
          type: number
          format: double
          description: Expected annual return (decimal, e.g., 0.15 = 15%)
        volatility:
          type: number
          format: double
          description: Expected annual volatility (decimal, e.g., 0.28 = 28%)
        sharpe_ratio:
          type: number
          format: double
          description: Sharpe ratio (return / volatility)
        optimization_method:
          type: string
          description: Optimization method used
        computed_at:
          type: string
          format: date-time
          description: When the optimization was computed
    PortfolioSummaryResponse:
      type: object
      required:
        - weights
      properties:
        weights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Optimal weight for each symbol
        expected_return:
          type: number
          format: double
          nullable: true
        volatility:
          type: number
          format: double
          nullable: true
        sharpe_ratio:
          type: number
          format: double
          nullable: true
    FrontierResponse:
      type: object
      required:
        - frontier
        - max_sharpe
        - min_variance
        - assets
      properties:
        frontier:
          type: array
          description: Points along the efficient frontier
          items:
            type: object
            properties:
              volatility:
                type: number
                format: double
              return:
                type: number
                format: double
              weights:
                type: array
                items:
                  type: number
        max_sharpe:
          type: object
          description: Maximum Sharpe ratio portfolio
          properties:
            volatility:
              type: number
              format: double
            return:
              type: number
              format: double
            weights:
              type: array
              items:
                type: number
        min_variance:
          type: object
          description: Minimum variance portfolio
          properties:
            volatility:
              type: number
              format: double
            return:
              type: number
              format: double
            weights:
              type: array
              items:
                type: number
        assets:
          type: array
          description: Individual asset risk-return positions
          items:
            type: object
            properties:
              symbol_index:
                type: integer
              volatility:
                type: number
                format: double
              return:
                type: number
                format: double
        capital_market_line:
          type: object
          description: CML from risk-free rate through tangency portfolio
          properties:
            x:
              type: array
              items:
                type: number
            y:
              type: array
              items:
                type: number
        risk_free_rate:
          type: number
          format: double
        symbols:
          type: array
          items:
            type: string
    BacktestResponse:
      type: object
      required:
        - windows
        - cumulative_values
        - metrics
        - config
      properties:
        windows:
          type: array
          description: Per-window optimization results
          items:
            type: object
            properties:
              window_id:
                type: integer
              train_start:
                type: string
                format: date
              train_end:
                type: string
                format: date
              test_start:
                type: string
                format: date
              test_end:
                type: string
                format: date
              weights:
                type: object
                additionalProperties:
                  type: number
              test_return:
                type: number
                format: double
        cumulative_values:
          type: object
          properties:
            dates:
              type: array
              items:
                type: string
                format: date
            strategy:
              type: array
              items:
                type: number
            equal_weight:
              type: array
              items:
                type: number
            btc_only:
              type: array
              items:
                type: number
        metrics:
          type: object
          description: Performance metrics for each strategy
          properties:
            strategy:
              $ref: '#/components/schemas/BacktestMetrics'
            equal_weight:
              $ref: '#/components/schemas/BacktestMetrics'
            btc_only:
              $ref: '#/components/schemas/BacktestMetrics'
        symbols:
          type: array
          items:
            type: string
        config:
          type: object
          properties:
            train_window:
              type: integer
            test_window:
              type: integer
            strategy:
              type: string
            risk_free_rate:
              type: number
              format: double
    BacktestMetrics:
      type: object
      properties:
        cumulative_return:
          type: number
          format: double
        annualized_return:
          type: number
          format: double
        max_drawdown:
          type: number
          format: double
        sharpe_ratio:
          type: number
          format: double
        calmar_ratio:
          type: number
          format: double
    MonteCarloResponse:
      type: object
      required:
        - percentiles
        - final_values
        - var_95
        - cvar_95
        - config
      properties:
        percentiles:
          type: object
          description: Percentile bands (p5, p25, p50, p75, p95)
          additionalProperties:
            type: array
            items:
              type: number
        final_values:
          type: array
          description: Final portfolio values for each simulation
          items:
            type: number
        var_95:
          type: number
          format: double
          description: Value at Risk (95%) as return from 1.0
        cvar_95:
          type: number
          format: double
          description: Conditional VaR (expected shortfall, 95%)
        symbols:
          type: array
          items:
            type: string
        config:
          type: object
          properties:
            n_simulations:
              type: integer
            n_days:
              type: integer
    RebalanceResponse:
      type: object
      required:
        - alerts
        - trades
        - summary
        - portfolio_key
      properties:
        alerts:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              current_weight:
                type: number
                format: double
              target_weight:
                type: number
                format: double
              drift:
                type: number
                format: double
              drift_pct:
                type: number
                format: double
              action:
                type: string
                enum: [BUY, SELL]
        trades:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              action:
                type: string
              amount_usd:
                type: number
                format: double
              weight_change:
                type: number
                format: double
        summary:
          type: object
          properties:
            total_drift:
              type: number
              format: double
            max_drift:
              type: number
              format: double
            n_alerts:
              type: integer
            n_symbols:
              type: integer
            drift_threshold:
              type: number
              format: double
            portfolio_value:
              type: number
              format: double
            needs_rebalance:
              type: boolean
        portfolio_key:
          type: string
    CombinedPortfolioResponse:
      type: object
      required:
        - weights
        - allocation
        - expected_return
        - volatility
        - sharpe_ratio
        - crypto_symbols
        - trad_symbols
        - crypto_metrics
        - trad_metrics
      properties:
        weights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Blended weights across all assets (sums to 1.0)
        allocation:
          type: object
          properties:
            crypto:
              type: number
              format: double
            traditional:
              type: number
              format: double
        expected_return:
          type: number
          format: double
        volatility:
          type: number
          format: double
        sharpe_ratio:
          type: number
          format: double
        crypto_symbols:
          type: array
          items:
            type: string
        trad_symbols:
          type: array
          items:
            type: string
        crypto_metrics:
          type: object
          description: Crypto sub-portfolio metrics
        trad_metrics:
          type: object
          description: Traditional sub-portfolio metrics
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (production only)

# Security disabled for MVP
# security:
#   - ApiKeyAuth: []